requirements.txt
streamlit
transformers
torch
newspaper3k